+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 Changed:
 - Filter out empty children with no `Attribute.TEXT` or `Attribute.RESOURCE_ID` in ViewHierarchy
+- Allow all `UiElement` options to be used for `ScrollOption`
 
 [0.11.0] - 2023-04-05